refactor: extract inline types, helpers, and constants from backend services
Move inline interfaces, utility functions, error classes, and constants
out of service files into dedicated modules:

- Shared types (TimeRangeConfig, SendEmailOptions) → @manifest/shared
- Backend-only types → per-module .types.ts files
- Crypto utilities → secret/crypto/encryption.utils.ts
- Schema helpers → node/schema/utils/
- MCP error class → mcp/errors/
- MCP template resolver → mcp/utils/
- App icons helper → app/utils/

Add code organization guidelines to CLAUDE.md to enforce this pattern.
