+++ packages/manifest/backend/src/flow-execution/flow-execution.service.ts
@@ -3,39 +3,17 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, LessThan } from 'typeorm';
 import { FlowExecutionEntity } from './flow-execution.entity';
 import type {
-  ExecutionStatus,
-  NodeExecutionData,
-  ExecutionErrorInfo,
   ExecutionListItem,
   ExecutionListResponse,
   FlowExecution,
 } from '@manifest/shared';
+import type {
+  CreateExecutionParams,
+  UpdateExecutionParams,
+  FindByFlowOptions,
+} from './flow-execution.types';
 
-export interface CreateExecutionParams {
-  flowId: string;
-  flowName: string;
-  flowToolName: string;
-  initialParams: Record<string, unknown>;
-  /** Whether this execution was triggered from preview chat (vs MCP) */
-  isPreview?: boolean;
-  /** Unique user fingerprint (hash of IP + User-Agent) for analytics */
-  userFingerprint?: string;
-}
-
-export interface UpdateExecutionParams {
-  status?: ExecutionStatus;
-  endedAt?: Date;
-  nodeExecutions?: NodeExecutionData[];
-  errorInfo?: ExecutionErrorInfo;
-}
-
-export interface FindByFlowOptions {
-  page?: number;
-  limit?: number;
-  status?: ExecutionStatus;
-  /** Filter by preview executions (true = only preview, false = only non-preview, undefined = all) */
-  isPreview?: boolean;
-}
+export type { CreateExecutionParams, UpdateExecutionParams, FindByFlowOptions } from './flow-execution.types';
 
 @Injectable()
 export class FlowExecutionService {
