+++ packages/manifest/backend/src/mcp/errors/mcp-inactive-tool.error.ts
@@ -1,0 +1,16 @@
+import { BadRequestException } from '@nestjs/common';
+
+/**
+ * MCP protocol error for inactive tools
+ * Error code -32602 is "Invalid params" per JSON-RPC 2.0 specification
+ */
+export class McpInactiveToolError extends BadRequestException {
+  readonly code = -32602;
+
+  constructor(toolName: string) {
+    super({
+      code: -32602,
+      message: `Tool '${toolName}' is not active and cannot be executed`,
+    });
+  }
+}
