+++ packages/manifest/backend/src/node/schema/schema.service.ts
@@ -31,6 +31,8 @@ import {
   type TestApiCallRequest,
   type TestApiCallResponse,
 } from '@manifest/shared';
+import { resolveTemplateVariables } from './utils/template-resolver.utils';
+import { parseAndValidateUrl } from './utils/url-validator.utils';
 
 /**
  * Service for schema-related operations.
@@ -243,7 +245,7 @@ export class SchemaService {
 
     // Resolve template variables in URL
     const { resolved: resolvedUrl, unresolvedVars: urlVars } =
-      this.resolveTemplateVariables(rawUrl, request.mockValues);
+      resolveTemplateVariables(rawUrl, request.mockValues);
 
     // Resolve template variables in headers
     const resolvedHeaders: Record<string, string> = {};
@@ -252,7 +254,7 @@ export class SchemaService {
     for (const header of rawHeaders) {
       if (header.key && header.value) {
         const { resolved, unresolvedVars } =
-          this.resolveTemplateVariables(header.value, request.mockValues);
+          resolveTemplateVariables(header.value, request.mockValues);
         resolvedHeaders[header.key] = resolved;
         headerUnresolvedVars.push(...unresolvedVars);
       }
@@ -270,7 +272,7 @@ export class SchemaService {
 
     // Validate URL format and block SSRF attacks
     // This is a security barrier - only validated URLs pass through
-    const urlValidation = this.parseAndValidateUrl(resolvedUrl);
+    const urlValidation = parseAndValidateUrl(resolvedUrl);
     if (!urlValidation.valid) {
       return {
         success: false,
@@ -393,69 +395,6 @@ export class SchemaService {
     }
   }
 
-  /**
-   * Sanitizes a mock value to prevent SSRF through URL injection.
-   * Blocks values that could be used to manipulate the target URL.
-   */
-  private sanitizeMockValue(value: unknown): string {
-    const str = String(value);
-
-    // Block values that look like full URLs (could override the base URL)
-    // This prevents injection of http://, https://, file://, etc.
-    if (/^[a-z][a-z0-9+.-]*:\/\//i.test(str)) {
-      return '[blocked-url]';
-    }
-
-    // Block protocol-relative URLs (//example.com)
-    if (str.startsWith('//')) {
-      return '[blocked-url]';
-    }
-
-    return str;
-  }
-
-  /**
-   * Resolve template variables in a string using mock values.
-   * Returns the resolved string and list of unresolved variable names.
-   * All mock values are sanitized to prevent SSRF through URL injection.
-   */
-  private resolveTemplateVariables(
-    template: string,
-    mockValues?: Record<string, unknown>
-  ): { resolved: string; unresolvedVars: string[] } {
-    const unresolvedVars: string[] = [];
-
-    // Match {{nodeSlug.path.to.field}} or {{nodeSlug}}
-    const resolved = template.replace(/\{\{([^}]+)\}\}/g, (match, varPath) => {
-      const parts = varPath.trim().split('.');
-      const rootKey = parts[0];
-
-      if (!mockValues || !(rootKey in mockValues)) {
-        unresolvedVars.push(varPath);
-        return match; // Keep original placeholder
-      }
-
-      // Navigate the path
-      let value: unknown = mockValues[rootKey];
-      for (let i = 1; i < parts.length && value != null; i++) {
-        if (typeof value === 'object' && value !== null) {
-          value = (value as Record<string, unknown>)[parts[i]];
-        } else {
-          value = undefined;
-        }
-      }
-
-      if (value === undefined || value === null) {
-        unresolvedVars.push(varPath);
-        return match;
-      }
-
-      // Sanitize the value to prevent SSRF through URL injection
-      return this.sanitizeMockValue(value);
-    });
-
-    return { resolved, unresolvedVars };
-  }
 
   // ==========================================================================
   // Connection Validation
@@ -787,85 +726,4 @@ export class SchemaService {
     return flow;
   }
 
-  /**
-   * Parses and validates a URL for SSRF vulnerabilities.
-   * Returns either a validated URL string (safe to fetch) or an error.
-   * This method acts as a security barrier - the returned URL is explicitly
-   * validated against SSRF attacks.
-   */
-  private parseAndValidateUrl(urlString: string): { valid: true; url: string } | { valid: false; error: string } {
-    let url: URL;
-    try {
-      url = new URL(urlString);
-    } catch {
-      return { valid: false, error: `Invalid URL format: ${urlString}` };
-    }
-
-    const ssrfError = this.checkSsrfVulnerability(url);
-    if (ssrfError) {
-      return { valid: false, error: ssrfError };
-    }
-
-    // Return a freshly constructed URL string from validated components
-    // This ensures only safe URLs pass through this security barrier
-    return { valid: true, url: url.href };
-  }
-
-  /**
-   * Checks a URL for SSRF vulnerabilities.
-   * Returns an error message if the URL is unsafe, or null if it's allowed.
-   */
-  private checkSsrfVulnerability(url: URL): string | null {
-    const hostname = url.hostname.toLowerCase();
-
-    // Block non-HTTP(S) protocols
-    if (url.protocol !== 'http:' && url.protocol !== 'https:') {
-      return `Protocol "${url.protocol}" is not allowed. Use HTTP or HTTPS.`;
-    }
-
-    // Block localhost
-    if (hostname === 'localhost' || hostname === '127.0.0.1' || hostname === '::1') {
-      return 'Requests to localhost are not allowed.';
-    }
-
-    // Block cloud metadata endpoints
-    if (hostname === '169.254.169.254' || hostname === 'metadata.google.internal') {
-      return 'Requests to cloud metadata endpoints are not allowed.';
-    }
-
-    // Check for private/internal IP ranges
-    const ipv4Match = hostname.match(/^(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})$/);
-    if (ipv4Match) {
-      const octets = ipv4Match.slice(1).map(Number);
-      const octet1 = octets[0];
-      const octet2 = octets[1];
-
-      // 10.0.0.0/8 - Private
-      if (octet1 === 10) {
-        return 'Requests to private IP addresses (10.x.x.x) are not allowed.';
-      }
-      // 172.16.0.0/12 - Private
-      if (octet1 === 172 && octet2 >= 16 && octet2 <= 31) {
-        return 'Requests to private IP addresses (172.16-31.x.x) are not allowed.';
-      }
-      // 192.168.0.0/16 - Private
-      if (octet1 === 192 && octet2 === 168) {
-        return 'Requests to private IP addresses (192.168.x.x) are not allowed.';
-      }
-      // 169.254.0.0/16 - Link-local
-      if (octet1 === 169 && octet2 === 254) {
-        return 'Requests to link-local addresses (169.254.x.x) are not allowed.';
-      }
-      // 127.0.0.0/8 - Loopback
-      if (octet1 === 127) {
-        return 'Requests to loopback addresses (127.x.x.x) are not allowed.';
-      }
-      // 0.0.0.0/8 - Current network
-      if (octet1 === 0) {
-        return 'Requests to 0.x.x.x addresses are not allowed.';
-      }
-    }
-
-    return null;
-  }
 }
